Counter

Description

Documentation

count :: IO [(Int, String)]

Count word frequencies and shows tuples containing word and it's frequency

convertToVec :: (Num a2, Eq a1) => [a1] -> [a1] -> [a2]

Convert string array into the binary vector, checking if it contains words included in patterns array