Parser

Description

Documentation

substring :: String -> String -> Bool

check if one string is substring of the second string

prefix :: String -> String -> Bool

check if one string is prefix of the second string

parse :: FilePath -> IO ()

parse an email to make it suitable for machine learnig algorithm

tooLower :: String -> String

parse string to lower cases

delete_punctuation_mark :: String -> String

remove punctuation mark

parseChar :: String -> String

parse string, by replacing some chars by their string substitiude like "xxx@xxx" -> "email"

deleteAppendix :: [Char] -> [Char]

remove appendix from email